refactor: remove binary meta-upgrade-09
To use it for upgrading, download it from an old release package, such
as:

https://github.com/datafuselabs/databend/releases/tag/v1.1.33-nightly
