--- src/binaries/meta-upgrade-09/Cargo.toml
@@ -1,33 +1,0 @@
-[package]
-name = "databend-meta-upgrade-09"
-description = "upgrade meta-service data to VER=24"
-version = { workspace = true }
-authors = { workspace = true }
-license = { workspace = true }
-publish = { workspace = true }
-edition = { workspace = true }
-
-[dependencies]
-# Workspace dependencies
-common-meta-app = { path = "../../meta/app" }
-common-meta-kvapi = { path = "../../meta/kvapi" }
-common-meta-raft-store = { path = "../../meta/raft-store" }
-common-meta-sled-store = { path = "../../meta/sled-store" }
-common-meta-types = { path = "../../meta/types" }
-common-proto-conv = { path = "../../meta/proto-conv" }
-common-protos = { path = "../../meta/protos" }
-common-tracing = { path = "../../common/tracing" }
-
-# Crates.io dependencies
-anyhow = { workspace = true }
-clap = { workspace = true }
-openraft = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-tokio = { workspace = true }
-
-[[bin]]
-name = "databend-meta-upgrade-09"
-path = "src/main.rs"
-doctest = false
-test = false
