+++ src/meta/process/src/kv_processor.rs
@@ -12,56 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#![allow(clippy::uninlined_format_args)]
-
-//! This program upgrades metadata written by databend-query before 0.9 to a databend-query-0.9 compatible format.
-//!
-//! It loads metadata from a `raft-dir` and upgrade TableMeta to new version and write them back.
-//! Both in raft-log data and in state-machine data will be converted.
-//!
-//! Usage:
-//!
-//! - Shut down all databend-meta processes.
-//!
-//! - Backup before proceeding: https://databend.rs/doc/deploy/metasrv/metasrv-backup-restore
-//!
-//! - Build it with `cargo build --bin databend-meta-upgrade-09`.
-//!
-//! - To view the current TableMeta version, print all TableMeta records with the following command:
-//!   It should display a list of TableMeta record.
-//!   You need to upgrade only if there is a `ver` that is lower than 24.
-//!
-//!    ```text
-//!    databend-meta-upgrade-09 --cmd print --raft-dir "<./your/raft-dir/>"
-//!    # output:
-//!    # TableMeta { ver: 23, ..
-//!    ```
-//!
-//! - Run it:
-//!
-//!   ```text
-//!   databend-meta-upgrade-09 --cmd upgrade --raft-dir "<./your/raft-dir/>"
-//!   ```
-//!
-//! - To assert upgrade has finished successfully, print all TableMeta records that are found in meta dir with the following command:
-//!   It should display a list of TableMeta record with a `ver` that is greater or equal 24.
-//!
-//!    ```text
-//!    databend-meta-upgrade-09 --cmd print --raft-dir "<./your/raft-dir/>"
-//!    # output:
-//!    # TableMeta { ver: 25, ..
-//!    # TableMeta { ver: 25, ..
-//!    ```
-
-mod rewrite;
-
 use anyhow::Error;
-use clap::Parser;
-use common_meta_app::schema::TableId;
-use common_meta_app::schema::TableMeta;
-use common_meta_kvapi::kvapi::Key;
 use common_meta_raft_store::key_spaces::RaftStoreEntry;
-use common_meta_sled_store::init_sled_db;
 use common_meta_types::txn_condition::Target;
 use common_meta_types::txn_op::Request;
 use common_meta_types::Cmd;
@@ -74,158 +26,14 @@ use common_meta_types::TxnOp;
 use common_meta_types::TxnPutRequest;
 use common_meta_types::TxnRequest;
 use common_meta_types::UpsertKV;
-use common_proto_conv::FromToProto;
-use common_protos::pb;
-use common_tracing::init_logging;
-use common_tracing::Config as LogConfig;
 use openraft::EntryPayload;
-use serde::Deserialize;
-use serde::Serialize;
 
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq, Parser)]
-#[clap(about, author)]
-pub struct Config {
-    #[clap(long, default_value = "INFO")]
-    pub log_level: String,
-
-    #[clap(long, default_value = "")]
-    pub cmd: String,
-
-    #[clap(flatten)]
-    pub raft_config: RaftConfig,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq, Parser)]
-#[clap(about, version, author)]
-#[serde(default)]
-pub struct RaftConfig {
-    /// The dir to store persisted meta state, including raft logs, state machine etc.
-    #[clap(long, default_value = "./_meta")]
-    #[serde(alias = "kvsrv_raft_dir")]
-    pub raft_dir: String,
-}
-
-impl Default for RaftConfig {
-    fn default() -> Self {
-        Self {
-            raft_dir: "./_meta".to_string(),
-        }
-    }
-}
-
-/// Usage:
-/// - To convert meta-data: `$0 --cmd upgrade --raft-dir ./_your_meta_dir/`:
-#[tokio::main]
-async fn main() -> anyhow::Result<()> {
-    let config = Config::parse();
-
-    let _guards = init_logging("databend-meta-upgrade-09", &LogConfig::default());
-
-    eprintln!();
-    eprintln!("███╗   ███╗███████╗████████╗ █████╗ ");
-    eprintln!("████╗ ████║██╔════╝╚══██╔══╝██╔══██╗");
-    eprintln!("██╔████╔██║█████╗     ██║   ███████║");
-    eprintln!("██║╚██╔╝██║██╔══╝     ██║   ██╔══██║");
-    eprintln!("██║ ╚═╝ ██║███████╗   ██║   ██║  ██║");
-    eprintln!("╚═╝     ╚═╝╚══════╝   ╚═╝   ╚═╝  ╚═╝");
-    eprintln!("     --- Upgrade to 0.9 ---         ");
-    eprintln!();
-
-    eprintln!("config: {}", pretty(&config)?);
-
-    match config.cmd.as_str() {
-        "print" => {
-            print_table_meta(&config)?;
-        }
-        "upgrade" => {
-            let p = GenericKVProcessor {
-                process_pb: conv_serialized_table_meta,
-            };
-            rewrite::rewrite(&config, |x| p.process(x))?;
-        }
-        _ => {
-            return Err(anyhow::anyhow!("invalid cmd: {:?}", config.cmd));
-        }
-    }
-
-    Ok(())
-}
-
-/// Print TableMeta in protobuf message format that are found in log or state machine.
-pub fn print_table_meta(config: &Config) -> anyhow::Result<()> {
-    let p = GenericKVProcessor {
-        process_pb: print_serialized_table_meta,
-    };
-
-    let raft_config = &config.raft_config;
-
-    init_sled_db(raft_config.raft_dir.clone());
-
-    for tree_iter_res in common_meta_sled_store::iter::<Vec<u8>>() {
-        let (_tree_name, item_iter) = tree_iter_res?;
-
-        for item_res in item_iter {
-            let (k, v) = item_res?;
-            let v1_ent = RaftStoreEntry::deserialize(&k, &v)?;
-
-            p.process(v1_ent)?;
-        }
-    }
-
-    Ok(())
-}
-
-fn conv_serialized_table_meta(key: &str, v: Vec<u8>) -> Result<Vec<u8>, anyhow::Error> {
-    // Only convert the key space `table-by-id`
-    if !key.starts_with(TableId::PREFIX) {
-        return Ok(v);
-    }
-
-    let p: pb::TableMeta = common_protos::prost::Message::decode(v.as_slice())?;
-    let v1: TableMeta = FromToProto::from_pb(p)?;
-
-    let p_latest = FromToProto::to_pb(&v1)?;
-
-    let mut buf = vec![];
-    common_protos::prost::Message::encode(&p_latest, &mut buf)?;
-    Ok(buf)
-}
-
-fn print_serialized_table_meta(key: &str, v: Vec<u8>) -> Result<Vec<u8>, anyhow::Error> {
-    if !key.starts_with(TableId::PREFIX) {
-        return Ok(v);
-    }
-
-    let p: pb::TableMeta = common_protos::prost::Message::decode(v.as_slice())?;
-
-    println!("{:?}", p);
-
-    Ok(v)
-}
-
-macro_rules! unwrap_or_return {
-    ($v: expr) => {
-        if let Some(x) = $v {
-            x
-        } else {
-            return Ok(None);
-        }
-    };
-}
-
-trait Process {
-    type Inp;
-    type Ret;
-
-    /// It returns None if nothing has to be done.
-    /// Otherwise it returns Some(converted_data).
-    fn process(&self, input: Self::Inp) -> Result<Option<Self::Ret>, anyhow::Error>;
-}
+use crate::process::Process;
 
 /// A processor that only processes GenericKV data.
 ///
 /// GenericKV data will be found in state-machine, logs that operates a GenericKV or a Txn.
-struct GenericKVProcessor<F>
+pub struct GenericKVProcessor<F>
 where F: Fn(&str, Vec<u8>) -> Result<Vec<u8>, anyhow::Error>
 {
     /// A callback to process serialized protobuf message
@@ -243,9 +51,22 @@ where F: Fn(&str, Vec<u8>) -> Result<Vec<u8>, anyhow::Error>
     }
 }
 
+macro_rules! unwrap_or_return {
+    ($v: expr) => {
+        if let Some(x) = $v {
+            x
+        } else {
+            return Ok(None);
+        }
+    };
+}
+
 impl<F> GenericKVProcessor<F>
 where F: Fn(&str, Vec<u8>) -> Result<Vec<u8>, anyhow::Error>
 {
+    pub fn new(process_pb: F) -> Self {
+        GenericKVProcessor { process_pb }
+    }
     /// Convert log and state machine record in meta-service.
     fn proc_raft_store_entry(
         &self,
@@ -401,9 +222,4 @@ where F: Fn(&str, Vec<u8>) -> Result<Vec<u8>, anyhow::Error>
 
         Ok(pr)
     }
-}
-
-fn pretty<T>(v: &T) -> Result<String, serde_json::Error>
-where T: Serialize {
-    serde_json::to_string_pretty(v)
 }
